to work

+++ MonteCarloTask.cpp
@@ -9,10 +9,11 @@ void* MonteCarloTask::operator()(void *_arg) {
     int numberOfIteration = *static_cast<int *>(arg);
     result = 0;
     for (int i = 0; i < numberOfIteration; i ++) {
-        srand(i * pthread_self()->__sig);
+        srand(static_cast<unsigned int>(i * pthread_self()->__sig));
         double x = static_cast<double>(rand()) / static_cast<double>(RAND_MAX);
         double y = static_cast<double>(rand()) / static_cast<double>(RAND_MAX);
-        if (x * x + y * y <= 1.0) result++;
+        if (x * x + y * y <= 1.0)
+            result++;
     }
     return NULL;
 }
